requirements.txt
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.31.0
requests==2.26.0
pytest==6.2.5
Flask==3.0.0
pandas==2.2.2
openpyxl==3.1.5